Integration +TDS+CVS

+++ include/integration_gyro.h
@@ -6,17 +6,17 @@ class integration_gyro
 public:
 
     integration_gyro();
-    integration_gyro(float** S,int n);
+    integration_gyro(double** S,int n);
     float* Integration_gyro_X(int n);
     float* Integration_gyro_Y(int n);
     float* Integration_gyro_Z(int n);
     void test();
 
 private:
-    float* VX;
-    float* VY;
-    float* VZ;
-    float* t;
+    double* VX;
+    double* VY;
+    double* VZ;
+    double* t;
 
 
 };
